modify read number fun

+++ algorithms trainings level 2/ABS function training/ABS function training.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 float ReadNumber()
 {
-	int Number;
+	float Number;
 
 	cout << "Please enter a number" << endl;
 	cin >> Number;
